added interface.c interface.h
