+++ include/interface.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <stdio.h>
+#include <mytype.h>
+#include <parser.h>
+#include <lexer.h>
+
+void match(token_t expected);
